lets

+++ main.h
@@ -2,7 +2,7 @@
 #define MAIN_H
 
 int _printf(const char *format, ...);
-int  print_number(long int n);
+int  print_number(int n);
 int print_string(char *s);
 int  myputchar(char c);
 int _strlen(char *s);
